+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - Listen also to the `change` and `onpaste` events, not only to the `onkeyup`.
 - Show the lowest verdict when the score is below zero.
+- New option to pass more input fields content to the zxcvbn library.
 
 ## 1.1.5